Split StreamUtil again, now extracting StreamCaptureUtils.
