+++ core/src/test/java/nl/nn/adapterframework/util/StreamUtilTest.java
@@ -3,7 +3,6 @@ package nl.nn.adapterframework.util;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-import java.io.BufferedOutputStream;
 import java.io.BufferedReader;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
@@ -23,15 +22,12 @@ import java.net.URL;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.Arrays;
 
 import org.junit.AfterClass;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
-import nl.nn.adapterframework.stream.Message;
-import nl.nn.adapterframework.util.StreamUtil.MarkCompensatingOutputStream;
 
 public class StreamUtilTest {
 
@@ -141,45 +137,6 @@ public class StreamUtilTest {
 		testStreamToByteArray("/StreamUtil/inUTF8withBOM.bin", false, UTF8_EXPECTED, true);
 	}
 
-	@Test
-	public void testCaptureWithMarkSupportedOutputStream() throws Exception {
-		URL input = ClassUtils.getResourceURL("/ForEachChildElementPipe/bulk2.xml");
-
-		int bufferSize = 2048;
-		Message message = new Message(input.openStream()); //non-repeatable
-		ByteArrayOutputStream boas = message.captureBinaryStream();
-		byte[] magic = message.getMagic(bufferSize);
-
-		message.asString(); //read twice after the magic has been fetched
-		message.asString();
-
-		byte[] capture = Arrays.copyOf(boas.toByteArray(), bufferSize); //it is possible more characters have been written to the captured stream
-		assertEquals(new String(magic), new String(capture));
-		assertEquals(message.asString(), Message.asMessage(input).asString()); //verify that the message output, after reading the magic, has not changed
-	}
-
-	@Test
-	public void testMarkSupportedOutputStream() throws Exception {
-		byte[] bytes = "A, B, C, D, E, F, G, H, I, J, K, L, M, N, O, P, Q, R, S, T, U, V, W, X, Y, Z".getBytes(); //76 chars
-		ByteArrayOutputStream boas = new ByteArrayOutputStream();
-		BufferedOutputStream bout = new BufferedOutputStream(boas, 16);
-		try (MarkCompensatingOutputStream out = new MarkCompensatingOutputStream(bout)) {
-
-			out.write(bytes, 9, 6);
-			out.flush();
-			assertEquals("D, E, ", new String(boas.toByteArray()));
-
-			for (int i = 0; i < bytes.length/4; i++) {
-				if(i%2==0) {
-					out.mark(3);
-				}
-				out.write(bytes, i*4, 4);
-			}
-		}
-
-		//D, E, prefix has already been written before the first skip has been called.
-		assertEquals("D, E, B, C, E, F, H, J, K, M, N, P, R, S, U, V, X, Z", new String(boas.toByteArray()));
-	}
 
 	private void writeToTestFile() throws IOException {
 		Writer w = new FileWriter(file.toString());
